Update M5600_hw1
I have updated the latex code for M5600 homework 1.  I am happy to be in 
a group with Dane Gollero and Ike Griss Salas.  They are trusting me to 
complete the latex for our group project and I will not let them down.  
I have updated this code to include exercises 9-14, with some additions 
still needed for exercise 12.  I will work on that after I collaborate 
with Dane.